Merge v0.8 to v1.0
Still no fix no nasty bug that won't display invalid MOVEMENT message, but I have a solution in mind. Instead of breaking up my command button and directional buttons, I am going to combine them to see if it will fix the error, and I will go from there. The matrix is complete, and I am currently still thinking of an idea for a secret room to be unlocked in the game. Wil also create a nice style sheet for final version, which will be a refresher from the dull border since project 2.